# This is a combination of 10 commits. # This is the 1st commit message:
adding jenkins release pipeline

# This is the commit message #2:

release process for metronome.  needs jenkins testing

# This is the commit message #3:

need protobuf installed and in path to build

# This is the commit message #4:

tagging is not useful right now

# This is the commit message #5:

need to be super apparently

# This is the commit message #6:

unable to get the cached protoc compiler

# This is the commit message #7:

last attempt.  it appears that the shell out that ammonite does loses the path

# This is the commit message #8:

going around ammonite for now

# This is the commit message #9:

my aliases do not work on jenkins

# This is the commit message #10:

crazy unable to build on the node