Execute dependent job when all dependencies finished. (#382)
Summary:
This the business logic of MIP 0008 "Jobs with Dependencies". We had an actor analog
to the `JobSpecSchedulerActor`. This new actor subscribes to all job run events and updates
an internal state. Once all parents have a successful run that is new than the child run we
trigger new job execution.

JIRA issues: DCOS_OSS-5974