fix: use the latest spec to run dependants
formerly we were using `initSpec` whenever we ran a dependany job. now
each dependencyActor also listens to `JobSpecUpdated` in order to update
its local copy of its jobSpec.

i'm really sure there is a more elegant solution to this. additionally
i'm not sure why we'd need an actor for managing dependencies at all.
but all that does not matter here, as we want to fix this by a given
deadline.

a major downside of this commit is that i'm not fluent enough in this
codebase to quickly assemble an according test.

closes MARATHON-8772